+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ Changelog
 Features:
 
 * Support for webapp2 via the `webargs.webapp2parser` module. Thanks :user:`Trii`.
+* Store argument name on ``RequiredArgMissingError``. Thanks :user:`stas`.
+* Allow error messages for required validation to be overriden. Thanks again :user:`stas`.
+
 
 0.12.0 (2015-03-22)
 *******************